# openapi-appengine.yaml
...
swagger: "2.0"
info:
  description: "API for Quran"
  title: "Quran API"
  version: "1.0.0"
host: "utopian-button-227405.appspot.com"
...
consumes:
  - "application/json"
produces:
  - "application/json"
schemes:
  - "https"
paths:
...
  "/v1/audio/ayah/{ayahId}":
    get:
      tags:
        - "Audio"
      description: "Get Audio by Ayah Id"
      operationId: "getAudioByAyahId"
      parameters:
        - in: path
          name: ayahId
          type: string
          required: true
          description: String ID of Ayah
        - in: query
          name: limit
          type: integer
          description: Limit the number of results
        - in: query
          name: offset
          type: integer
          description: Start getting result from specific offset usually for next page results
      produces:
        - "application/json"
      responses:
        200:
          description: "Successful operation to get Audio"
          schema:
            $ref: "#/definitions/audio"
...
  "/v1/ayah/{ayahId}":
    get:
      tags:
        - "Ayah"
      description: "Get Ayah by Id"
      operationId: "getAyahById"
      parameters:
        - in: path
          name: ayahId
          type: string
          required: true
          description: String ID of Ayah
        - in: query
          name: parts
          type: string
          description: Comma seperated additional parts possible value are Translation,Surah,Edition,Arabic_Audio,Translation_Audio,Image. You can set "All" if you want to get all parts.
        - in: query
          name: edition
          type: string
          description: String id of edition
      produces:
        - "application/json"
      responses:
        200:
          description: "Successful operation to get Ayah"
          schema:
            $ref: "#/definitions/ayah"
  "/v1/ayah/surah/{surahId}":
    get:
      tags:
        - "Ayah"
      description: "Get Ayah by Surah Id"
      operationId: "getAyahBySurahId"
      parameters:
        - in: path
          name: surahId
          type: string
          required: true
          description: String ID of Surah
        - in: query
          name: limit
          type: integer
          description: Limit the number of results
        - in: query
          name: offset
          type: integer
          description: Start getting result from specific offset usually for next page results
        - in: query
          name: parts
          type: string
          description: Comma seperated additional parts possible value are Translation,Surah,Edition,Arabic_Audio,Translation_Audio,Image. You can set "All" if you want to get all parts.
        - in: query
          name: edition
          type: string
          description: String id of edition
      produces:
        - "application/json"
      responses:
        200:
          description: "Successful operation to get Ayah"
          schema:
            $ref: "#/definitions/ayah"
  "/v1/ayah/number/{number}":
    get:
      tags:
        - "Ayah"
      description: "Get Ayah by Surah number"
      operationId: "getAyahByNumber"
      parameters:
        - in: path
          name: number
          type: integer
          required: true
          description: Number of surah
        - in: query
          name: parts
          type: string
          description: Comma seperated additional parts possible value are Translation,Surah,Edition,Arabic_Audio,Translation_Audio,Image. You can set "All" if you want to get all parts.
        - in: query
          name: edition
          type: string
          description: String id of edition
      produces:
        - "application/json"
      responses:
        200:
          description: "Successful operation to get Ayah"
          schema:
            $ref: "#/definitions/ayah"
  "/v1/ayah/number_in_surah/{number}":
    get:
      tags:
        - "Ayah"
      description: "Get Ayah by Number if surah"
      operationId: "getAyahByNumberInSurah"
      parameters:
        - in: path
          name: number
          type: integer
          required: true
          description: Number of surah
        - in: query
          name: parts
          type: string
          description: Comma seperated additional parts possible value are Translation,Surah,Edition,Arabic_Audio,Translation_Audio,Image. You can set "All" if you want to get all parts.
        - in: query
          name: edition
          type: string
          description: String id of edition
      produces:
        - "application/json"
      responses:
        200:
          description: "Successful operation to get Ayah"
          schema:
            $ref: "#/definitions/ayah"
  "/v1/ayah/juz/{juz}":
    get:
      tags:
        - "Ayah"
      description: "Get Ayah by Juz"
      operationId: "getAyahByJuz"
      parameters:
        - in: path
          name: juz
          type: integer
          required: true
          description: Number of juz
        - in: query
          name: limit
          type: integer
          description: Limit the number of results
        - in: query
          name: offset
          type: integer
          description: Start getting result from specific offset usually for next page results
        - in: query
          name: parts
          type: string
          description: Comma seperated additional parts possible value are Translation,Surah,Edition,Arabic_Audio,Translation_Audio,Image. You can set "All" if you want to get all parts.
        - in: query
          name: edition
          type: string
          description: String id of edition
      produces:
        - "application/json"
      responses:
        200:
          description: "Successful operation to get Ayah"
          schema:
            $ref: "#/definitions/ayah"
  "/v1/ayah/manzil/{manzil}":
    get:
      tags:
        - "Ayah"
      description: "Get Ayah by manzil"
      operationId: "getAyahByManzil"
      parameters:
        - in: path
          name: manzil
          type: integer
          required: true
          description: Number of manzil
        - in: query
          name: limit
          type: integer
          description: Limit the number of results
        - in: query
          name: offset
          type: integer
          description: Start getting result from specific offset usually for next page results
        - in: query
          name: parts
          type: string
          description: Comma seperated additional parts possible value are Translation,Surah,Edition,Arabic_Audio,Translation_Audio,Image. You can set "All" if you want to get all parts.
        - in: query
          name: edition
          type: string
          description: String id of edition
      produces:
        - "application/json"
      responses:
        200:
          description: "Successful operation to get Ayah"
          schema:
            $ref: "#/definitions/ayah"
  "/v1/ayah/ruku/{ruku}":
    get:
      tags:
        - "Ayah"
      description: "Get Ayah by ruku number"
      operationId: "getAyahByRuku"
      parameters:
        - in: path
          name: ruku
          type: integer
          required: true
          description: Number of ruku
        - in: query
          name: limit
          type: integer
          description: Limit the number of results
        - in: query
          name: offset
          type: integer
          description: Start getting result from specific offset usually for next page results
        - in: query
          name: parts
          type: string
          description: Comma seperated additional parts possible value are Translation,Surah,Edition,Arabic_Audio,Translation_Audio,Image. You can set "All" if you want to get all parts.
        - in: query
          name: edition
          type: string
          description: String id of edition
      produces:
        - "application/json"
      responses:
        200:
          description: "Successful operation to get Ayah"
          schema:
            $ref: "#/definitions/ayah"
  "/v1/ayah/hizb_qurater/{hizbQuarter}":
    get:
      tags:
        - "Ayah"
      description: "Get Ayah by hizb quarter"
      operationId: "getAyahByHizbQuarter"
      parameters:
        - in: path
          name: hizbQuarter
          type: integer
          required: true
          description: Number of hizbQuarter
        - in: query
          name: limit
          type: integer
          description: Limit the number of results
        - in: query
          name: offset
          type: integer
          description: Start getting result from specific offset usually for next page results
        - in: query
          name: parts
          type: string
          description: Comma seperated additional parts possible value are Translation,Surah,Edition,Arabic_Audio,Translation_Audio,Image. You can set "All" if you want to get all parts.
        - in: query
          name: edition
          type: string
          description: String id of edition
      produces:
        - "application/json"
      responses:
        200:
          description: "Successful operation to get Ayah"
          schema:
            $ref: "#/definitions/ayah"
  "/v1/ayah/sajda/{sajda}":
    get:
      tags:
        - "Ayah"
      description: "Get Ayah by sajda"
      operationId: "getAyahBySajda"
      parameters:
        - in: path
          name: sajda
          type: boolean
          required: true
          description: Is sajda Ayah or not
        - in: query
          name: limit
          type: integer
          description: Limit the number of results
        - in: query
          name: offset
          type: integer
          description: Start getting result from specific offset usually for next page results
        - in: query
          name: parts
          type: string
          description: Comma seperated additional parts possible value are Translation,Surah,Edition,Arabic_Audio,Translation_Audio,Image. You can set "All" if you want to get all parts.
        - in: query
          name: edition
          type: string
          description: String id of edition
      produces:
        - "application/json"
      responses:
        200:
          description: "Successful operation to get Ayah"
          schema:
            $ref: "#/definitions/ayah"
...
  "/v1/surah/{surahId}":
    get:
      tags:
        - "Surah"
      description: "Get Surah by ID"
      operationId: "getSurahById"
      parameters:
        - in: path
          name: surahId
          type: string
          required: true
          description: String ID of surah
      produces:
        - "application/json"
      responses:
        200:
          description: "Successful operation to get Surah"
          schema:
            $ref: "#/definitions/surah"
...
  "/v1/translation/ayah/{ayahId}":
    get:
      tags:
        - "Translation"
      description: "Get Translation by ayahId"
      operationId: "getTranslationAyahId"
      parameters:
        - in: path
          name: ayahId
          type: string
          required: true
          description: String ID of Ayah
        - in: query
          name: limit
          type: integer
          description: Limit the number of results
        - in: query
          name: offset
          type: integer
          description: Start getting result from specific offset usually for next page results
      produces:
        - "application/json"
      responses:
        200:
          description: "Successful operation to get Translation"
          schema:
            $ref: "#/definitions/translation"
...
  "/v1/translation/filter/{ayahId}/{editionId}":
    get:
      tags:
        - "Translation"
      description: "Filter Translation by ayahId and EditionId"
      operationId: "filterTranslation"
      parameters:
        - in: path
          name: ayahId
          type: string
          required: true
          description: String ID of Ayah
        - in: path
          name: editionId
          type: string
          required: true
          description: String ID of Edition
      produces:
        - "application/json"
      responses:
        200:
          description: "Successful operation to get Translation"
          schema:
            $ref: "#/definitions/translation"
definitions:
  audio:
    type: "object"
    properties:
      id:
        type: string
      ayahId:
        type: string
      ayahNumber:
        type: integer
      editionId:
        type: string
      type:
        type: string
        description: Type can be Audio or Translation
      link:
        type: string
        description: Link of the Audio file
  ayah:
    type: object
    properties:
      id:
        type: string
      surahId:
        type: string
      number:
        type: integer
      numberInSurah:
        type: integer
      juz:
        type: integer
      manzil:
        type: integer
      ruku:
        type: integer
      hizbQuarter:
        type: integer
      sajda:
        type: boolean
      arabic:
        type: string
...
  surah:
    type: object
    properties:
      id:
        type: string
      number:
        type: integer
      name:
        type: string
      englishName:
        type: string
      englishNameTranslation:
        type: string
      numberOfAyahs:
        type: integer
      revelationType:
        type: string
  translation:
    type: object
    properties:
      id:
        type: string
      ayahId:
        type: string
      ayahNumber:
        type: integer
      editionId:
        type: string
      text:
        type: string

# Does not require an API key.
security: []